+++ demo/demo/urls.py
@@ -13,10 +13,20 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
+import django
 from django.contrib import admin
-from django.urls import path, include
 
-urlpatterns = [
-    path('', include('post.urls')),
-    path('admin/', admin.site.urls),
-]
+if django.VERSION >= (2, 0):
+    from django.urls import path, include
+
+    urlpatterns = [
+        path('', include('post.urls')),
+        path('admin/', admin.site.urls),
+    ]
+else:
+    from django.conf.urls import url, include
+
+    urlpatterns = [
+        url(r'^', include('post.urls')),
+        url(r'^admin/', admin.site.urls),
+    ]