+++ popup_field/views.py
@@ -1,75 +1,19 @@
+import django
 from django.utils.decorators import classonlymethod
 from django.views.generic import CreateView, UpdateView, DeleteView
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.http.response import JsonResponse
 from django.template.response import TemplateResponse
 from django.core.exceptions import ImproperlyConfigured
-from django.urls import path, include
-
 from .fields import ForeignKeyWidget, ManyToManyWidget
 
-
-class AttributeThunk(object):
-    """
-    Class thunks various attributes expected by Django generic CRUD views as
-    properties of the parent viewset class instance. This allows us to
-    normalize all CRUD view attributes as ViewSet properties and/or methods.
-    """
-
-    def __init__(self, viewset, *args, **kwargs):
-        self._viewset = viewset()
-        self.raise_exception = self._viewset.raise_exception
-        # allow viewset methods to access view
-        self._viewset.view = self
-        # self._viewset.template_name_create = self._viewset.template_name_create
-        # self._viewset.template_name_update = self._viewset.template_name_update
-        super(AttributeThunk, self).__init__(*args, **kwargs)
-
-    @property
-    def model(self):
-        return self._viewset.model
-
-    @property
-    def fields(self):
-        return self._viewset.fields
-
-    def get_form_class(self):
-        if hasattr(self._viewset, 'form_class'):
-            return self._viewset.form_class
-        return super(AttributeThunk, self).get_form_class()
-
-    def get_form_kwargs(self):
-        kwargs = super(AttributeThunk, self).get_form_kwargs()
-        kwargs.update(self._viewset.get_form_kwargs())
-        return kwargs
-
-    def get_permission_required(self):
-        actions = {
-            'PopupCreateView': 'create',
-            'PopupUpdateView': 'update',
-            'PopupDeleteView': 'delete'
-        }
-        return self._viewset.get_permission_required(actions[self.__class__.__name__])
+if django.VERSION >= (2, 0):
+    from django.urls import path, include
+else:
+    from django.conf.urls import url, include
 
 
-class TemplateNameMixin(object):
-    """
-    Get attr template_name_create to PopupCreateView
-    Get attr template_name_update to PopupUpdateView
-    """
-
-    def get_template_names(self):
-        templates = super(TemplateNameMixin, self).get_template_names()
-        if self.__class__.__name__ == 'PopupCreateView':
-            template_attr_name = 'template_name_create'
-        else:
-            template_attr_name = 'template_name_update'
-        if hasattr(self._viewset, template_attr_name):
-            templates.insert(0, getattr(self._viewset, template_attr_name))
-        return templates
-
-
-class PopupCreateView(AttributeThunk, TemplateNameMixin, PermissionRequiredMixin, CreateView):
+class PopupCreateView(PermissionRequiredMixin, CreateView):
 
     def get_context_data(self, **kwargs):
         if 'to_field' in self.request.GET:
@@ -85,7 +29,7 @@ class PopupCreateView(AttributeThunk, TemplateNameMixin, PermissionRequiredMixin
         return TemplateResponse(self.request, 'popup/success.html', context=context)
 
 
-class PopupUpdateView(AttributeThunk, TemplateNameMixin, PermissionRequiredMixin, UpdateView):
+class PopupUpdateView(PermissionRequiredMixin, UpdateView):
     slug_field = 'id'
     context_object_name = 'popup'
 
@@ -103,7 +47,7 @@ class PopupUpdateView(AttributeThunk, TemplateNameMixin, PermissionRequiredMixin
         return TemplateResponse(self.request, 'popup/success.html', context=context)
 
 
-class PopupDeleteView(AttributeThunk, PermissionRequiredMixin, DeleteView):
+class PopupDeleteView(PermissionRequiredMixin, DeleteView):
     slug_field = 'id'
 
     def delete(self, request, *args, **kwargs):
@@ -118,8 +62,9 @@ class PopupDeleteView(AttributeThunk, PermissionRequiredMixin, DeleteView):
 
 class PopupCRUDViewSet(object):
     model = None
-    fields = ()
     form_class = None
+    template_name_create = None
+    template_name_update = None
     """
     permissions_required = {
         'create': ('post.add_category',),
@@ -131,81 +76,73 @@ class PopupCRUDViewSet(object):
     permissions_required = {}
 
     @classonlymethod
-    def _generate_view(cls, crud_view_class, **initkwargs):
-        def view(request, *args, **kwargs):
-            initkwargs['request'] = request
-            view = crud_view_class(cls, **initkwargs)
-            if hasattr(view, 'get') and not hasattr(view, 'head'):
-                view.head = view.get
-            view.request = request
-            view.args = args
-            view.kwargs = kwargs
-            return view.dispatch(request, *args, **kwargs)
-
-        view.view_class = crud_view_class
-        view.view_initkwargs = initkwargs
-
-        # take name and docstring from class
-        # update_wrapper(view, crud_view_class, updated=())
-
-        # and possible attributes set by decorators
-        # like csrf_exempt from dispatch
-        # update_wrapper(view, crud_view_class.dispatch, assigned=())
-        return view
-
-    @classonlymethod
-    def create(cls, **initkwargs):
-        """Returns the create view that can be specified as the second argument
+    def create(cls):
+        """
+        Returns the create view that can be specified as the second argument
         to url() in urls.py.
         """
-        return cls._generate_view(PopupCreateView, **initkwargs)
+
+        class NewPopupCreateView(PopupCreateView):
+            model = cls.model
+            form_class = cls.form_class
+            template_name = cls.template_name_create
+            permission_required = cls.get_permission_required('create')
+
+        return NewPopupCreateView
 
     @classonlymethod
-    def update(cls, **initkwargs):
-        """Returns the update view that can be specified as the second argument
+    def update(cls):
+        """
+        Returns the update view that can be specified as the second argument
         to url() in urls.py.
         """
-        return cls._generate_view(PopupUpdateView, **initkwargs)
+
+        class NewPopupUpdateView(PopupUpdateView):
+            model = cls.model
+            form_class = cls.form_class
+            template_name = cls.template_name_update
+            permission_required = cls.get_permission_required('update')
+
+        return NewPopupUpdateView
 
     @classonlymethod
-    def delete(cls, **initkwargs):
-        """Returns the delete view that can be specified as the second argument
+    def delete(cls):
+        """
+        Returns the delete view that can be specified as the second argument
         to url() in urls.py.
         """
-        return cls._generate_view(PopupDeleteView, **initkwargs)
 
-    def get_form_kwargs(self):
-        """
-        For Create and Update views, this method allows passing custom arguments
-        to the form class constructor. The return value from this method is
-        combined with the default form constructor ``**kwargs`` before it is
-        passed to the form class' ``__init__()`` routine's ``**kwargs``.
+        class PopupDeleteViewView(PopupDeleteView):
+            model = cls.model
+            form_class = cls.form_class
+            permission_required = cls.get_permission_required('delete')
 
-        Since Django CBVs use kwargs ``initial`` & ``instance``, be careful
-        when using these, unless of course, you want to override the objects
-        provided by these keys.
-        """
-        return {}
+        return PopupDeleteViewView
 
     @classonlymethod
     def urls(cls):
         """
         generate url and url_name for create、update and delete view
-        default url_name is classname_
+        default url_name is classname_name
         """
         class_name = cls.model.__name__.lower()
-        return path('{}/'.format(class_name), include([
-            path('popup/', cls.create(), name='{}_popup_create'.format(class_name)),
-            path('popup/<int:pk>/', cls.update(), name='{}_popup_update'.format(class_name)),
-            path('popup/delete/<int:pk>/', cls.delete(), name='{}_popup_delete'.format(class_name)),
-        ]))
+        if django.VERSION >= (2, 0):
+            return path('{}/'.format(class_name), include([
+                path('popup/', cls.create().as_view(), name='{}_popup_create'.format(class_name)),
+                path('popup/<int:pk>/', cls.update().as_view(), name='{}_popup_update'.format(class_name)),
+                path('popup/delete/<int:pk>/', cls.delete().as_view(), name='{}_popup_delete'.format(class_name)),
+            ]))
+        else:
+            return url(r'^{}/'.format(class_name), include([
+                url(r'^popup/$', cls.create().as_view(), name='{}_popup_create'.format(class_name)),
+                url(r'^popup/(?P<pk>\d+)/$', cls.update().as_view(), name='{}_popup_update'.format(class_name)),
+                url(r'^popup/delete/(?P<pk>\d+)/$', cls.delete().as_view(), name='{}_popup_delete'.format(class_name)),
+            ]))
 
     @classonlymethod
     def get_fk_popup_field(cls, *args, **kwargs):
         """
         generate fk field related to class wait popup crud
-        :param request:
-        :return:
         """
         class_name = cls.model.__name__.lower()
         class_verbose_name = cls.model._meta.verbose_name
@@ -217,8 +154,6 @@ class PopupCRUDViewSet(object):
     def get_m2m_popup_field(cls, *args, **kwargs):
         """
         generate m2m field related to class wait popup crud
-        :param model_class:
-        :return:
         """
         class_name = cls.model.__name__.lower()
         class_verbose_name = cls.model._meta.verbose_name
@@ -226,14 +161,10 @@ class PopupCRUDViewSet(object):
         kwargs['permissions_required'] = cls.permissions_required
         return ManyToManyWidget('{}_popup_create'.format(class_name), *args, **kwargs)
 
-    def get_permission_required(self, action):
+    @classonlymethod
+    def get_permission_required(cls, action):
         """
-        Return the permission required for the CRUD operation specified in op.
+        Return the permission required for the CRUD operation specified in action.
         Default implementation returns the value of one
-        ``{list|create|detail|update|delete}_permission_required`` class attributes.
-        Overriding this allows you to return dynamically computed permissions.
-
-        :param op: The CRUD operation code. One of
-            ``{'list'|'create'|'detail'|'update'|'delete'}``.
         """
-        return self.permissions_required.get(action, [])
+        return cls.permissions_required.get(action, [])