+++ HISTORY.md
@@ -6,3 +6,7 @@
 
 ### v 0.1.2
 - allow custom template_name for popup_field
+
+### v 0.1.3
+- compatible with django 1.x
+- change viewset achieve logic